python-cli-framework/requirements.txt
PyYAML>=6.0
packaging>=21.0
